stashing working code
